further linter issue fixes
